Fix warning related to poinless comparison of unsigned int

+++ src/ekat/util/ekat_string_utils.cpp
@@ -124,7 +124,7 @@ bool valid_nested_list_format (const std::string& str)
   // We verified the string starts with '['.
   size_t start = 1;
   char last_match = open;
-  size_t num_open = 1;
+  int num_open = 1;
   size_t pos = str.find_first_of(separators,start);
 
   while (pos!=npos) {
